requirements.txt
discord==2.3.2
SQLAlchemy==2.0.30
python-dotenv==1.0.1
feedparser==6.0.11